requirements.txt
streamlit
yfinance
pandas
numpy
langchain-google-genai
langchain-core
python-dotenv